requirements.txt
 
pandas
sentence_transformers
faiss-cpu
streamlit
